Move ExternalProviders to a file
