+++ src/sentry/api/endpoints/project_details.py
@@ -37,10 +37,10 @@ from sentry.models import (
     ProjectStatus,
     ProjectTeam,
 )
-from sentry.models.integration import ExternalProviders
 from sentry.notifications.legacy_mappings import get_option_value_from_boolean
 from sentry.notifications.types import NotificationSettingTypes
 from sentry.tasks.deletion import delete_project
+from sentry.types.integrations import ExternalProviders
 from sentry.utils import json
 from sentry.utils.compat import filter
