+++ src/sentry/api/endpoints/user_notification_fine_tuning.py
@@ -6,12 +6,12 @@ from sentry.api.bases.user import UserEndpoint
 from sentry.api.serializers import serialize
 from sentry.api.serializers.models import UserNotificationsSerializer
 from sentry.models import NotificationSetting, Project, UserEmail, UserOption
-from sentry.models.integration import ExternalProviders
 from sentry.notifications.legacy_mappings import (
     get_option_value_from_int,
     get_type_from_fine_tuning_key,
 )
 from sentry.notifications.types import FineTuningAPIKey
+from sentry.types.integrations import ExternalProviders
 
 INVALID_EMAIL_MSG = (
     "Invalid email value(s) provided. Email values must be verified emails for the given user."
