+++ src/sentry/api/serializers/models/user_notifications.py
@@ -3,12 +3,12 @@ from typing import Iterable
 
 from sentry.api.serializers import Serializer
 from sentry.models import NotificationSetting, UserOption
-from sentry.models.integration import ExternalProviders
 from sentry.notifications.legacy_mappings import (
     get_type_from_fine_tuning_key,
     map_notification_settings_to_legacy,
 )
 from sentry.notifications.types import FineTuningAPIKey, NotificationScopeType
+from sentry.types.integrations import ExternalProviders
 
 
 def handle_legacy(notification_type: FineTuningAPIKey, users: Iterable) -> Iterable:
