+++ src/sentry/models/notificationsetting.py
@@ -7,7 +7,6 @@ from sentry.db.models import (
     Model,
     sane_repr,
 )
-from sentry.models.integration import ExternalProviders, get_provider_name
 from sentry.notifications.manager import NotificationsManager
 from sentry.notifications.types import (
     NotificationScopeType,
@@ -17,6 +16,7 @@ from sentry.notifications.types import (
     get_notification_setting_type_name,
     get_notification_setting_value_name,
 )
+from sentry.types.integrations import ExternalProviders, get_provider_name
 
 
 class NotificationSetting(Model):
