+++ src/sentry/web/frontend/accounts.py
@@ -14,10 +14,10 @@ from django.views.decorators.csrf import csrf_protect
 from django.views.decorators.http import require_http_methods
 
 from sentry.models import Authenticator, LostPasswordHash, NotificationSetting, Project, UserEmail
-from sentry.models.integration import ExternalProviders
 from sentry.notifications.types import NotificationSettingOptionValues, NotificationSettingTypes
 from sentry.security import capture_security_activity
 from sentry.signals import email_verified
+from sentry.types.integrations import ExternalProviders
 from sentry.utils import auth
 from sentry.web.decorators import login_required, set_referrer_policy, signed_auth_required
 from sentry.web.forms.accounts import ChangePasswordRecoverForm, RecoverPasswordForm
