+++ tests/sentry/models/test_groupsubscription.py
@@ -1,13 +1,13 @@
 import functools
 
 from sentry.models import GroupSubscription, NotificationSetting
-from sentry.models.integration import ExternalProviders
 from sentry.notifications.types import (
     GroupSubscriptionReason,
     NotificationSettingOptionValues,
     NotificationSettingTypes,
 )
 from sentry.testutils import TestCase
+from sentry.types.integrations import ExternalProviders
 
 
 class SubscribeTest(TestCase):
